requirements.txt
streamlit
pymongo
requests
google-generativeai
tqdm
numpy
python-dotenv
gradio
cerebras-cloud-sdk
tiktoken
fastapi
uvicorn
pymongo
python-dotenv
